refactor to use widgets
