+++ AppFramework/src/main/java/com/balch/android/app/framework/MetadataUtils.java
@@ -32,13 +32,13 @@ import com.balch.android.app.framework.domain.EditState;
 import com.balch.android.app.framework.domain.ViewHint;
 import com.balch.android.app.framework.domain.annotations.ColumnEdit;
 import com.balch.android.app.framework.domain.annotations.ColumnNew;
-import com.balch.android.app.framework.domain.widgets.EditLayout;
-import com.balch.android.app.framework.domain.widgets.BoolEditLayout;
-import com.balch.android.app.framework.domain.widgets.EnumEditLayout;
-import com.balch.android.app.framework.domain.widgets.MoneyEditLayout;
-import com.balch.android.app.framework.domain.widgets.NumberEditLayout;
-import com.balch.android.app.framework.domain.widgets.StringEditLayout;
-import com.balch.android.app.framework.domain.widgets.UnsupportedEditLayout;
+import com.balch.android.app.framework.domain.widget.EditLayout;
+import com.balch.android.app.framework.domain.widget.BoolEditLayout;
+import com.balch.android.app.framework.domain.widget.EnumEditLayout;
+import com.balch.android.app.framework.domain.widget.MoneyEditLayout;
+import com.balch.android.app.framework.domain.widget.NumberEditLayout;
+import com.balch.android.app.framework.domain.widget.StringEditLayout;
+import com.balch.android.app.framework.domain.widget.UnsupportedEditLayout;
 import com.balch.android.app.framework.domain.DomainObject;
 import com.balch.android.app.framework.types.Money;
