+++ AppFramework/src/main/java/com/balch/android/app/framework/domain/ColumnDescriptor.java
@@ -24,7 +24,7 @@ package com.balch.android.app.framework.domain;
 
 import android.support.annotation.NonNull;
 
-import com.balch.android.app.framework.domain.widgets.EditLayout;
+import com.balch.android.app.framework.domain.widget.EditLayout;
 
 import java.lang.reflect.Field;
