+++ AppFramework/src/main/java/com/balch/android/app/framework/domain/widget/ControlMapper.java
@@ -20,7 +20,7 @@
  * Copyright (C) 2014
  */
 
-package com.balch.android.app.framework.domain.widgets;
+package com.balch.android.app.framework.domain.widget;
 
 public interface ControlMapper {
     ControlMap getControlMap();
