+++ MockTradeApp/src/main/java/com/balch/mocktrade/order/OrderEditController.java
@@ -29,9 +29,9 @@ import android.view.ViewGroup;
 import com.balch.android.app.framework.domain.ColumnDescriptor;
 import com.balch.android.app.framework.domain.ExternalController;
 import com.balch.android.app.framework.domain.ValidatorException;
-import com.balch.android.app.framework.domain.widgets.ControlMap;
-import com.balch.android.app.framework.domain.widgets.EditLayout;
-import com.balch.android.app.framework.domain.widgets.EnumEditLayout;
+import com.balch.android.app.framework.domain.widget.ControlMap;
+import com.balch.android.app.framework.domain.widget.EditLayout;
+import com.balch.android.app.framework.domain.widget.EnumEditLayout;
 import com.balch.android.app.framework.model.ModelFactory;
 import com.balch.android.app.framework.types.Money;
 import com.balch.mocktrade.R;
