+++ MockTradeApp/src/main/java/com/balch/mocktrade/order/QuantityPriceLayout.java
@@ -43,8 +43,8 @@ import com.balch.android.app.framework.domain.ColumnDescriptor;
 import com.balch.android.app.framework.domain.EditState;
 import com.balch.android.app.framework.domain.ValidatorException;
 import com.balch.android.app.framework.domain.ViewHint;
-import com.balch.android.app.framework.domain.widgets.ControlMapper;
-import com.balch.android.app.framework.domain.widgets.EditLayout;
+import com.balch.android.app.framework.domain.widget.ControlMapper;
+import com.balch.android.app.framework.domain.widget.EditLayout;
 import com.balch.android.app.framework.types.Money;
 import com.balch.mocktrade.R;
 import com.balch.mocktrade.account.Account;
