+++ MockTradeApp/src/main/java/com/balch/mocktrade/portfolio/DailyGraphView.java
@@ -1,121 +1,26 @@
 package com.balch.mocktrade.portfolio;
 
-import android.animation.Animator;
-import android.animation.ObjectAnimator;
-import android.animation.ValueAnimator;
 import android.content.Context;
 import android.content.res.TypedArray;
-import android.graphics.BlurMaskFilter;
 import android.graphics.Canvas;
-import android.graphics.Matrix;
-import android.graphics.Paint;
-import android.graphics.PointF;
-import android.graphics.RectF;
-import android.graphics.Shader;
-import android.graphics.SweepGradient;
 import android.os.Build;
 import android.util.AttributeSet;
 import android.view.View;
-import android.view.ViewGroup;
-import android.widget.Scroller;
 
 import com.balch.mocktrade.R;
 
-import java.util.ArrayList;
-import java.util.List;
 
-public class DailyGraphView extends ViewGroup {
+public class DailyGraphView extends View {
     private static final String TAG = DailyGraphView.class.getSimpleName();
 
-
-    private List<Item> mData = new ArrayList<Item>();
-
-    private float mTotal = 0.0f;
-
-    private RectF mPieBounds = new RectF();
-
-    private Paint mPiePaint;
-    private Paint mTextPaint;
-    private Paint mShadowPaint;
-
-    private float mTextX = 0.0f;
-    private float mTextY = 0.0f;
-    private float mTextWidth = 0.0f;
-    private float mTextHeight = 0.0f;
-    private int mTextPos = TEXTPOS_LEFT;
-
-    private float mHighlightStrength = 1.15f;
-
-    private int mTextColor;
-    private PieView mPieView;
-    private Scroller mScroller;
-    private ValueAnimator mScrollAnimator;
-    private PointerView mPointerView;
-
-    // the index of the current item.
-    private ObjectAnimator mAutoCenterAnimator;
-    private RectF mShadowBounds = new RectF();
-
-    /**
-     * Draw text to the left of the pie chart
-     */
-    public static final int TEXTPOS_LEFT = 0;
-
-    /**
-     * Draw text to the right of the pie chart
-     */
-    public static final int TEXTPOS_RIGHT = 1;
-
-    /**
-     * The initial fling velocity is divided by this amount.
-     */
-    public static final int FLING_VELOCITY_DOWNSCALE = 4;
-
-    /**
-     *
-     */
-    public static final int AUTOCENTER_ANIM_DURATION = 250;
-
-    /**
-     * Interface definition for a callback to be invoked when the current
-     * item changes.
-     */
-    public interface OnCurrentItemChangedListener {
-        void OnCurrentItemChanged(DailyGraphView source, int currentItem);
-    }
-
-    /**
-     * Class constructor taking only a context. Use this constructor to create
-     * {@link DailyGraphView} objects from your own code.
-     *
-     * @param context
-     */
     public DailyGraphView(Context context) {
         super(context);
-        init();
+        initialize();
     }
 
-    /**
-     * Class constructor taking a context and an attribute set. This constructor
-     * is used by the layout engine to construct a {@link DailyGraphView} from a set of
-     * XML attributes.
-     *
-     * @param context
-     * @param attrs   An attribute set which can contain attributes from
-     *                 as well as attributes inherited
-     *                from {@link android.view.View}.
-     */
     public DailyGraphView(Context context, AttributeSet attrs) {
         super(context, attrs);
 
-        // attrs contains the raw values for the XML attributes
-        // that were specified in the layout, which don't include
-        // attributes set by styles or themes, and which may have
-        // unresolved references. Call obtainStyledAttributes()
-        // to get the final values for each attribute.
-        //
-        // This call uses R.styleable.PieChart, which is an array of
-        // the custom attributes that were declared in attrs.xml.
         TypedArray a = context.getTheme().obtainStyledAttributes(
                 attrs,
                 R.styleable.DailyGraphView,
@@ -123,350 +28,30 @@ public class DailyGraphView extends ViewGroup {
         );
 
         try {
-            // Retrieve the values from the TypedArray and store into
-            // fields of this class.
-            //
-            // The R.styleable.PieChart_* constants represent the index for
-            // each custom attribute in the R.styleable.PieChart array.
+/*
             mTextY = a.getDimension(R.styleable.DailyGraphView_labelY, 0.0f);
             mTextWidth = a.getDimension(R.styleable.DailyGraphView_labelWidth, 0.0f);
             mTextHeight = a.getDimension(R.styleable.DailyGraphView_labelHeight, 0.0f);
             mTextPos = a.getInteger(R.styleable.DailyGraphView_labelPosition, 0);
             mTextColor = a.getColor(R.styleable.DailyGraphView_labelColor, 0xff000000);
+*/
         } finally {
             // release the TypedArray so that it can be reused.
             a.recycle();
         }
 
-        init();
+        initialize();
     }
 
-    /**
-     * Returns the Y position of the label text, in pixels.
-     *
-     * @return The Y position of the label text, in pixels.
-     */
-    public float getTextY() {
-        return mTextY;
-    }
-
-    /**
-     * Set the Y position of the label text, in pixels.
-     *
-     * @param textY the Y position of the label text, in pixels.
-     */
-    public void setTextY(float textY) {
-        mTextY = textY;
-        invalidate();
-    }
-
-    /**
-     * Returns the width reserved for label text, in pixels.
-     *
-     * @return The width reserved for label text, in pixels.
-     */
-    public float getTextWidth() {
-        return mTextWidth;
-    }
-
-    /**
-     * Set the width of the area reserved for label text. This width is constant; it does not
-     * change based on the actual width of the label as the label text changes.
-     *
-     * @param textWidth The width reserved for label text, in pixels.
-     */
-    public void setTextWidth(float textWidth) {
-        mTextWidth = textWidth;
-        invalidate();
-    }
-
-    /**
-     * Returns the height of the label font, in pixels.
-     *
-     * @return The height of the label font, in pixels.
-     */
-    public float getTextHeight() {
-        return mTextHeight;
-    }
-
-    /**
-     * Set the height of the label font, in pixels.
-     *
-     * @param textHeight The height of the label font, in pixels.
-     */
-    public void setTextHeight(float textHeight) {
-        mTextHeight = textHeight;
-        invalidate();
-    }
-
-    /**
-     * Returns a value that specifies whether the label text is to the right
-     * or the left of the pie chart graphic.
-     *
-     * @return One of TEXTPOS_LEFT or TEXTPOS_RIGHT.
-     */
-    public int getTextPos() {
-        return mTextPos;
-    }
-
-    /**
-     * Set a value that specifies whether the label text is to the right
-     * or the left of the pie chart graphic.
-     *
-     * @param textPos TEXTPOS_LEFT to draw the text to the left of the graphic,
-     *                or TEXTPOS_RIGHT to draw the text to the right of the graphic.
-     */
-    public void setTextPos(int textPos) {
-        if (textPos != TEXTPOS_LEFT && textPos != TEXTPOS_RIGHT) {
-            throw new IllegalArgumentException(
-                    "TextPos must be one of TEXTPOS_LEFT or TEXTPOS_RIGHT");
-        }
-        mTextPos = textPos;
-        invalidate();
-    }
-
-
-
-    @Override
-    protected void onLayout(boolean changed, int l, int t, int r, int b) {
-        // Do nothing. Do not call the superclass method--that would start a layout pass
-        // on this view's children. PieChart lays out its children in onSizeChanged().
-    }
 
 
     @Override
     protected void onDraw(Canvas canvas) {
         super.onDraw(canvas);
 
-        // Draw the shadow
-        canvas.drawOval(mShadowBounds, mShadowPaint);
-
-        // Draw the label text
-        canvas.drawText("Daily Performance", mTextX, mTextY, mTextPaint);
-
     }
 
-
-    //
-    // Measurement functions. This example uses a simple heuristic: it assumes that
-    // the pie chart should be at least as wide as its label.
-    //
-    @Override
-    protected int getSuggestedMinimumWidth() {
-        return (int) mTextWidth * 2;
-    }
-
-    @Override
-    protected int getSuggestedMinimumHeight() {
-        return (int) mTextWidth;
-    }
-
-    @Override
-    protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
-        // Try for a width based on our minimum
-        int minw = getPaddingLeft() + getPaddingRight() + getSuggestedMinimumWidth();
-
-        int w = Math.max(minw, MeasureSpec.getSize(widthMeasureSpec));
-
-        // Whatever the width ends up being, ask for a height that would let the pie
-        // get as big as it can
-        int minh = (w - (int) mTextWidth) + getPaddingBottom() + getPaddingTop();
-        int h = Math.min(MeasureSpec.getSize(heightMeasureSpec), minh);
-
-        setMeasuredDimension(w, h);
-    }
-
-    @Override
-    protected void onSizeChanged(int w, int h, int oldw, int oldh) {
-        super.onSizeChanged(w, h, oldw, oldh);
-
-        //
-        // Set dimensions for text, pie chart, etc
-        //
-        // Account for padding
-        float xpad = (float) (getPaddingLeft() + getPaddingRight());
-        float ypad = (float) (getPaddingTop() + getPaddingBottom());
-
-        // Account for the label
-        xpad += mTextWidth;
-
-        float ww = (float) w - xpad;
-        float hh = (float) h - ypad;
-
-        // Figure out how big we can make the pie.
-        float diameter = Math.min(ww, hh);
-        mPieBounds = new RectF(
-                0.0f,
-                0.0f,
-                diameter,
-                diameter);
-        mPieBounds.offsetTo(getPaddingLeft(), getPaddingTop());
-
-        // Make adjustments based on text position
-        if (mTextPos == TEXTPOS_LEFT) {
-            mTextPaint.setTextAlign(Paint.Align.RIGHT);
-            mPieBounds.offset(mTextWidth, 0.0f);
-            mTextX = mPieBounds.left;
-
-        } else {
-            mTextPaint.setTextAlign(Paint.Align.LEFT);
-            mTextX = mPieBounds.right;
-        }
-
-        mShadowBounds = new RectF(
-                mPieBounds.left + 10,
-                mPieBounds.bottom + 10,
-                mPieBounds.right - 10,
-                mPieBounds.bottom + 20);
-
-        // Lay out the child view that actually draws the pie.
-        mPieView.layout((int) mPieBounds.left,
-                (int) mPieBounds.top,
-                (int) mPieBounds.right,
-                (int) mPieBounds.bottom);
-        mPieView.setPivot(mPieBounds.width() / 2, mPieBounds.height() / 2);
-
-        mPointerView.layout(0, 0, w, h);
-        onDataChanged();
-    }
-
-    /**
-     * Do all of the recalculations needed when the data array changes.
-     */
-    private void onDataChanged() {
-        // When the data changes, we have to recalculate
-        // all of the angles.
-        int currentAngle = 0;
-        for (Item it : mData) {
-            it.mStartAngle = currentAngle;
-            it.mEndAngle = (int) ((float) currentAngle + it.mValue * 360.0f / mTotal);
-            currentAngle = it.mEndAngle;
-
-
-            // Recalculate the gradient shaders. There are
-            // three values in this gradient, even though only
-            // two are necessary, in order to work around
-            // a bug in certain versions of the graphics engine
-            // that expects at least three values if the
-            // positions array is non-null.
-            //
-            it.mShader = new SweepGradient(
-                    mPieBounds.width() / 2.0f,
-                    mPieBounds.height() / 2.0f,
-                    new int[]{
-                            it.mHighlight,
-                            it.mHighlight,
-                            it.mColor,
-                            it.mColor,
-                    },
-                    new float[]{
-                            0,
-                            (float) (360 - it.mEndAngle) / 360.0f,
-                            (float) (360 - it.mStartAngle) / 360.0f,
-                            1.0f
-                    }
-            );
-        }
-//        calcCurrentItem();
-        onScrollFinished();
-    }
-
-    /**
-     * Initialize the control. This code is in a separate method so that it can be
-     * called from both constructors.
-     */
-    private void init() {
-        // Force the background to software rendering because otherwise the Blur
-        // filter won't work.
-        setLayerToSW(this);
-
-        // Set up the paint for the label text
-        mTextPaint = new Paint(Paint.ANTI_ALIAS_FLAG);
-        mTextPaint.setColor(mTextColor);
-        if (mTextHeight == 0) {
-            mTextHeight = mTextPaint.getTextSize();
-        } else {
-            mTextPaint.setTextSize(mTextHeight);
-        }
-
-        // Set up the paint for the pie slices
-        mPiePaint = new Paint(Paint.ANTI_ALIAS_FLAG);
-        mPiePaint.setStyle(Paint.Style.FILL);
-        mPiePaint.setTextSize(mTextHeight);
-
-        // Set up the paint for the shadow
-        mShadowPaint = new Paint(0);
-        mShadowPaint.setColor(0xff101010);
-        mShadowPaint.setMaskFilter(new BlurMaskFilter(8, BlurMaskFilter.Blur.NORMAL));
-
-        // Add a child view to draw the pie. Putting this in a child view
-        // makes it possible to draw it on a separate hardware layer that rotates
-        // independently
-        mPieView = new PieView(getContext());
-        addView(mPieView);
-        mPieView.rotateTo(0);
-
-        // The pointer doesn't need hardware acceleration, but in order to show up
-        // in front of the pie it also needs to be on a separate view.
-        mPointerView = new PointerView(getContext());
-        addView(mPointerView);
-
-        // Set up an animator to animate the PieRotation property. This is used to
-        // correct the pie's orientation after the user lets go of it.
-        if (Build.VERSION.SDK_INT >= 11) {
-            mAutoCenterAnimator = ObjectAnimator.ofInt(DailyGraphView.this, "PieRotation", 0);
-
-            // Add a listener to hook the onAnimationEnd event so that we can do
-            // some cleanup when the pie stops moving.
-            mAutoCenterAnimator.addListener(new Animator.AnimatorListener() {
-                public void onAnimationStart(Animator animator) {
-                }
-
-                public void onAnimationEnd(Animator animator) {
-                    mPieView.decelerate();
-                }
-
-                public void onAnimationCancel(Animator animator) {
-                }
-
-                public void onAnimationRepeat(Animator animator) {
-                }
-            });
-        }
-
-
-        // Create a Scroller to handle the fling gesture.
-        if (Build.VERSION.SDK_INT < 11) {
-            mScroller = new Scroller(getContext());
-        } else {
-            mScroller = new Scroller(getContext(), null, true);
-        }
-        // The scroller doesn't have any built-in animation functions--it just supplies
-        // values when we ask it to. So we have to have a way to call it every frame
-        // until the fling ends. This code (ab)uses a ValueAnimator object to generate
-        // a callback on every animation frame. We don't use the animated value at all.
-        if (Build.VERSION.SDK_INT >= 11) {
-            mScrollAnimator = ValueAnimator.ofFloat(0, 1);
-            mScrollAnimator.addUpdateListener(new ValueAnimator.AnimatorUpdateListener() {
-                public void onAnimationUpdate(ValueAnimator valueAnimator) {
-                    tickScrollAnimation();
-                }
-            });
-        }
-
-
-    }
-
-    private void tickScrollAnimation() {
-        if (!mScroller.isFinished()) {
-            mScroller.computeScrollOffset();
-//            setPieRotation(mScroller.getCurrY());
-        } else {
-            if (Build.VERSION.SDK_INT >= 11) {
-                mScrollAnimator.cancel();
-            }
-            onScrollFinished();
-        }
+    private void initialize() {
     }
 
     private void setLayerToSW(View v) {
@@ -480,172 +65,5 @@ public class DailyGraphView extends ViewGroup {
             setLayerType(View.LAYER_TYPE_HARDWARE, null);
         }
     }
-
-    /**
-     * Force a stop to all pie motion. Called when the user taps during a fling.
-     */
-    private void stopScrolling() {
-        mScroller.forceFinished(true);
-        if (Build.VERSION.SDK_INT >= 11) {
-            mAutoCenterAnimator.cancel();
-        }
-
-        onScrollFinished();
-    }
-
-    /**
-     * Called when the user finishes a scroll action.
-     */
-    private void onScrollFinished() {
-        mPieView.decelerate();
-    }
-
-
-    /**
-     * Internal child class that draws the pie chart onto a separate hardware layer
-     * when necessary.
-     */
-    private class PieView extends View {
-        // Used for SDK < 11
-        private float mRotation = 0;
-        private Matrix mTransform = new Matrix();
-        private PointF mPivot = new PointF();
-
-        /**
-         * Construct a PieView
-         *
-         * @param context
-         */
-        public PieView(Context context) {
-            super(context);
-        }
-
-        /**
-         * Enable hardware acceleration (consumes memory)
-         */
-        public void accelerate() {
-            setLayerToHW(this);
-        }
-
-        /**
-         * Disable hardware acceleration (releases memory)
-         */
-        public void decelerate() {
-            setLayerToSW(this);
-        }
-
-        @Override
-        protected void onDraw(Canvas canvas) {
-            super.onDraw(canvas);
-
-            if (Build.VERSION.SDK_INT < 11) {
-                mTransform.set(canvas.getMatrix());
-                mTransform.preRotate(mRotation, mPivot.x, mPivot.y);
-                canvas.setMatrix(mTransform);
-            }
-
-            for (Item it : mData) {
-                mPiePaint.setShader(it.mShader);
-                canvas.drawArc(mBounds,
-                        360 - it.mEndAngle,
-                        it.mEndAngle - it.mStartAngle,
-                        true, mPiePaint);
-            }
-        }
-
-        @Override
-        protected void onSizeChanged(int w, int h, int oldw, int oldh) {
-            mBounds = new RectF(0, 0, w, h);
-        }
-
-        RectF mBounds;
-
-        public void rotateTo(float pieRotation) {
-            mRotation = pieRotation;
-            if (Build.VERSION.SDK_INT >= 11) {
-                setRotation(pieRotation);
-            } else {
-                invalidate();
-            }
-        }
-
-        public void setPivot(float x, float y) {
-            mPivot.x = x;
-            mPivot.y = y;
-            if (Build.VERSION.SDK_INT >= 11) {
-                setPivotX(x);
-                setPivotY(y);
-            } else {
-                invalidate();
-            }
-        }
-    }
-
-    /**
-     * View that draws the pointer on top of the pie chart
-     */
-    private class PointerView extends View {
-
-        /**
-         * Construct a PointerView object
-         *
-         * @param context
-         */
-        public PointerView(Context context) {
-            super(context);
-        }
-
-        @Override
-        protected void onDraw(Canvas canvas) {
-//            canvas.drawLine(mTextX, mPointerY, mPointerX, mPointerY, mTextPaint);
-//            canvas.drawCircle(mPointerX, mPointerY, mPointerRadius, mTextPaint);
-        }
-    }
-
-    /**
-     * Maintains the state for a data item.
-     */
-    private class Item {
-        public String mLabel;
-        public float mValue;
-        public int mColor;
-
-        // computed values
-        public int mStartAngle;
-        public int mEndAngle;
-
-        public int mHighlight;
-        public Shader mShader;
-    }
-
-
-    private boolean isAnimationRunning() {
-        return !mScroller.isFinished() || (Build.VERSION.SDK_INT >= 11 && mAutoCenterAnimator.isRunning());
-    }
-
-    /**
-     * Helper method for translating (x,y) scroll vectors into scalar rotation of the pie.
-     *
-     * @param dx The x component of the current scroll vector.
-     * @param dy The y component of the current scroll vector.
-     * @param x  The x position of the current touch, relative to the pie center.
-     * @param y  The y position of the current touch, relative to the pie center.
-     * @return The scalar representing the change in angular position for this scroll.
-     */
-    private static float vectorToScalarScroll(float dx, float dy, float x, float y) {
-        // get the length of the vector
-        float l = (float) Math.sqrt(dx * dx + dy * dy);
-
-        // decide if the scalar should be negative or positive by finding
-        // the dot product of the vector perpendicular to (x,y).
-        float crossX = -y;
-        float crossY = x;
-
-        float dot = (crossX * dx + crossY * dy);
-        float sign = Math.signum(dot);
-
-        return l * sign;
-    }
-
 
 }
